dont wait for prometheus namespace teardown
